Move tests from `src/query/mod.rs` to `src/query/sequence.rs`
